Add bootstrap.py to get rid of ansible installation step
Brain is a muscle, and it gets tired over time. While it
is not a problem to set everything manually, setting up
a workstation could be just a single step on the long
road to some goal. It may be a few steps or a hundred,
so wasting an extra minute on each step makes it 100
minutes. And that also the reason to keep everything
reliable and simple.